Feature/refatoracao controllers

+++ MedicalCare/Controllers/PacienteController.cs
@@ -1,9 +1,11 @@
 ﻿using MedicalCare.DTO;
 using MedicalCare.Interfaces;
 using MedicalCare.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
+using System.Security.Cryptography;
 
 namespace MedicalCare.Controllers
 {
@@ -12,20 +14,32 @@ namespace MedicalCare.Controllers
     public class PacienteController : ControllerBase
     {
         private readonly IPacienteService _pacienteService;
+        private readonly ILogService _logService;
         private readonly IRepository<PacienteModel> _repository;
 
-        public PacienteController(IPacienteService pacienteService, IRepository<PacienteModel> repository)
+        public PacienteController(IPacienteService pacienteService, IRepository<PacienteModel> repository, ILogService logService)
         {
             _pacienteService = pacienteService;
             _repository = repository;
+            _logService = logService;
         }
 
 
+
+        [Authorize(Roles = "Administrador, Médico, Enfermeiro")]
         [HttpPost]
         public ActionResult<PacienteGetDto> Post([FromBody] PacienteCreateDto pacienteCreate)
         {
             try
             {
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
+                {
+                    return BadRequest("Usuário inativo no sistema");
+                }
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+                int id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
                 bool verificaCpfEmail = _repository.GetByCpfEmail(pacienteCreate.Cpf, pacienteCreate.Email);
                 if (verificaCpfEmail)
                 {
@@ -33,6 +47,14 @@ namespace MedicalCare.Controllers
                 }
                 pacienteCreate.StatusDoSistema = true;
                 PacienteGetDto pacienteGet = _pacienteService.CreatePaciente(pacienteCreate);
+
+                LogModel logModel = new LogModel
+                {
+                    Descricao = $"{tipo} {nome}, de Id {id}, cadastrou o paciente {pacienteCreate.NomeCompleto} no sistema.",
+                    Dominio = "Paciente-cadastro."
+                };
+                _logService.CreateLog(logModel);
+
                 return Created("Paciente salvo com sucesso.", pacienteGet);
 
             }
@@ -43,12 +65,32 @@ namespace MedicalCare.Controllers
             }
         }
 
+
+
+        [Authorize(Roles = "Administrador, Médico, Enfermeiro")]
         [HttpGet]
         public ActionResult<IEnumerable<PacienteGetDto>> Get()
         {
             try
             {
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
+                {
+                    return BadRequest("Usuário inativo no sistema");
+                }
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+                int _id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
+
                 IEnumerable<PacienteGetDto> pacientes = _pacienteService.GetAllPacientes();
+
+                LogModel logModel = new LogModel
+                {
+                    Descricao = $"{tipo} {nome}, de Id {_id}, listou todos os pacientes do sistema.",
+                    Dominio = "Paciente-getAll."
+                };
+                _logService.CreateLog(logModel);
+
                 return Ok(pacientes);
             }
             catch (Exception)
@@ -58,16 +100,35 @@ namespace MedicalCare.Controllers
             }
         }
 
+
+
+        [Authorize(Roles = "Administrador, Médico, Enfermeiro")]
         [HttpGet("{id}")]
         public ActionResult<PacienteGetDto> Get([FromRoute] int id)
         {
             try
             {
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
+                {
+                    return BadRequest("Usuário inativo no sistema");
+                }
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+                int _id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
+
                 PacienteGetDto pacienteGet = _pacienteService.GetById(id);
                 if (pacienteGet == null)
                 {
                     return NoContent();
                 }
+
+                LogModel logModel = new LogModel
+                {
+                    Descricao = $"{tipo} {nome}, de Id {_id}, listou o paciente {pacienteGet.NomeCompleto} no sistema.",
+                    Dominio = "Paciente-getById."
+                };
+                _logService.CreateLog(logModel);
                 return Ok(pacienteGet);
             }
             catch (Exception)
@@ -77,17 +138,37 @@ namespace MedicalCare.Controllers
             }
         }
 
+
+
+        [Authorize(Roles = "Administrador, Médico, Enfermeiro")]
         [HttpPut("{id}")]
         public ActionResult<PacienteGetDto> Update([FromRoute] int id, [FromBody] PacienteUpdateDto pacienteUpdate)
         {
             try
             {
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
+                {
+                    return BadRequest("Usuário inativo no sistema");
+                }
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+                int _id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
+
                 PacienteGetDto verificaSeExiste = _pacienteService.GetById(id);
                 if (verificaSeExiste == null)
                 {
                     return NoContent();
                 }
                 PacienteGetDto pacienteGet = _pacienteService.UpdatePaciente(pacienteUpdate, id);
+
+                LogModel logModel = new LogModel
+                {
+                    Descricao = $"{tipo} {nome}, de Id {_id}, atualizou o paciente {pacienteUpdate.NomeCompleto} no sistema.",
+                    Dominio = "Paciente-update."
+                };
+                _logService.CreateLog(logModel);
+
                 return Ok(pacienteGet);
 
             }
@@ -98,15 +179,35 @@ namespace MedicalCare.Controllers
             }
         }
 
+
+
+        [Authorize(Roles = "Administrador, Médico, Enfermeiro")]
         [HttpDelete("{id}")]
         public ActionResult Delete([FromRoute] int id)
         {
             try
             {
-                bool remocao = _pacienteService.DeletePaciente(id);
-                if (remocao)
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
                 {
-                return Accepted();
+                    return BadRequest("Usuário inativo no sistema");
+                }
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+                int _id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
+                PacienteGetDto pacienteGet = _pacienteService.GetById(id);
+
+                if (pacienteGet != null)
+                {
+
+                    _pacienteService.DeletePaciente(id);
+                    LogModel logModel = new LogModel
+                    {
+                        Descricao = $"{tipo} {nome}, de Id {_id}, excluiu o paciente {pacienteGet.NomeCompleto}.",
+                        Dominio = "Paciente-delete."
+                    };
+                    _logService.CreateLog(logModel);
+                    return Accepted();
                 }
                 return NoContent();
